backend/requirements.txt
Django==5.2.13
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-filter==25.2
Pillow==11.2.1
psycopg2-binary==2.9.11
python-dotenv==1.1.0
drf-spectacular==0.29.0
firebase-admin==7.4.0
